@@ -16,20 +16,29 @@
             "homepage": "https://github.com/magnusnordlander"
         }
     ],
-    "homepage": "http://www.php-cache.com/en/latest/",
+    "homepage": "https://www.php-cache.com/en/latest/",
     "require": {
-        "php": ">=7.4",
-        "psr/cache": "^1.0 || ^2.0",
-        "psr/simple-cache": "^1.0"
+        "php": ">=8.1",
+        "psr/cache": "^3.0",
+        "psr/simple-cache": "^3.0"
     },
     "require-dev": {
-        "cache/integration-tests": "^0.17",
-        "mockery/mockery": "^1.0",
-        "phpunit/phpunit": "^7.5.20 || ^9.5.10",
+        "cache/integration-tests": "3.x-dev",
+        "phpunit/phpunit": "^10.0",
         "symfony/cache": "^3.2"
     },
     "provide": {
         "psr/simple-cache-implementation": "^1.0"
+    },
+    "repositories": {
+        "cache/integration-tests": {
+            "type": "git",
+            "url": "https://github.com/Sweetchuck/cache-integration-tests.git"
+        },
+        "tedivm/stash": {
+            "type": "github",
+            "url": "https://github.com/Sweetchuck/tedivm-stash.git"
+        }
     },
     "minimum-stability": "dev",
     "prefer-stable": true,
@@ -40,6 +49,13 @@
         "exclude-from-classmap": [
             "/Tests/"
         ]
+    },
+    "config": {
+        "optimize-autoloader": true,
+        "preferred-install": {
+            "*": "dist"
+        },
+        "sort-packages": true
     },
     "extra": {
         "branch-alias": {
