@@ -1,6 +1,9 @@
 <?php
 
-/*
+declare(strict_types = 1);
+
+/**
+ * @file
  * This file is part of php-cache organization.
  *
  * (c) 2015 Aaron Scherer <aequasi@gmail.com>, Tobias Nyholm <tobias.nyholm@gmail.com>
@@ -12,25 +15,21 @@
 namespace Cache\Bridge\SimpleCache\Tests;
 
 use Cache\Bridge\SimpleCache\SimpleCacheBridge;
-use Mockery as m;
 use PHPUnit\Framework\TestCase;
 use Psr\Cache\CacheItemInterface;
 use Psr\Cache\CacheItemPoolInterface;
 
 class SimpleCacheBridgeTest extends TestCase
 {
-    /**
-     * @type SimpleCacheBridge
-     */
-    private $bridge;
+    private SimpleCacheBridge $bridge;
 
     /**
-     * @type m\MockInterface|CacheItemPoolInterface
+     * @phpstan-var \PHPUnit\Framework\MockObject\MockObject&\Psr\Cache\CacheItemPoolInterface
      */
-    private $mock;
+    private $poolMock;
 
     /**
-     * @type m\MockInterface|CacheItemInterface
+     * @phpstan-var \PHPUnit\Framework\MockObject\MockObject&\Psr\Cache\CacheItemInterface
      */
     private $itemMock;
 
@@ -38,62 +37,115 @@ class SimpleCacheBridgeTest extends TestCase
     {
         parent::setUp();
 
-        $this->mock = m::mock(CacheItemPoolInterface::class);
+        $this->poolMock = $this
+            ->getMockBuilder(CacheItemPoolInterface::class)
+            ->getMock();
 
-        $this->bridge = new SimpleCacheBridge($this->mock);
+        $this->bridge = new SimpleCacheBridge($this->poolMock);
 
-        $this->itemMock = m::mock(CacheItemInterface::class);
+        $this->itemMock = $this
+            ->getMockBuilder(CacheItemInterface::class)
+            ->getMock();
     }
 
-    public function testConstructor()
+    public function testConstructor(): void
     {
-        $this->assertInstanceOf(SimpleCacheBridge::class, $this->bridge);
+        static::assertInstanceOf(SimpleCacheBridge::class, $this->bridge);
     }
 
-    public function testFetch()
+    public function testFetch(): void
     {
-        $this->itemMock->shouldReceive('isHit')->times(1)->andReturn(true);
-        $this->itemMock->shouldReceive('get')->times(1)->andReturn('some_value');
+        $this
+            ->itemMock
+            ->expects(static::once())
+            ->method('isHit')
+            ->willReturn(true);
+        $this
+            ->itemMock
+            ->expects(static::once())
+            ->method('get')
+            ->willReturn('some_value');
 
-        $this->mock->shouldReceive('getItem')->withArgs(['some_item'])->andReturn($this->itemMock);
+        $this
+            ->poolMock
+            ->expects(static::once())
+            ->method('getItem')
+            ->willReturn($this->itemMock);
 
-        $this->assertEquals('some_value', $this->bridge->get('some_item'));
+        static::assertSame('some_value', $this->bridge->get('some_item'));
     }
 
-    public function testFetchMiss()
+    public function testFetchMiss(): void
     {
-        $this->itemMock->shouldReceive('isHit')->times(1)->andReturn(false);
+        $this
+            ->itemMock
+            ->expects(static::once())
+            ->method('isHit')
+            ->willReturn(false);
 
-        $this->mock->shouldReceive('getItem')->withArgs(['no_item'])->andReturn($this->itemMock);
+        $this
+            ->poolMock
+            ->expects(static::once())
+            ->method('getItem')
+            ->with('no_item')
+            ->willReturn($this->itemMock);
 
-        $this->assertFalse($this->bridge->get('no_item', false));
+        static::assertFalse($this->bridge->get('no_item', false));
     }
 
-    public function testContains()
+    public function testContains(): void
     {
-        $this->mock->shouldReceive('hasItem')->withArgs(['no_item'])->andReturn(false);
-        $this->mock->shouldReceive('hasItem')->withArgs(['some_item'])->andReturn(true);
+        $this
+            ->poolMock
+            ->method('hasItem')
+            ->willReturnMap([
+                ['no_item', false],
+                ['some_item', true],
+            ]);
 
-        $this->assertFalse($this->bridge->has('no_item'));
-        $this->assertTrue($this->bridge->has('some_item'));
+        static::assertFalse($this->bridge->has('no_item'));
+        static::assertTrue($this->bridge->has('some_item'));
     }
 
-    public function testSave()
+    public function testSave(): void
     {
-        $this->itemMock->shouldReceive('set')->twice()->with('dummy_data');
-        $this->itemMock->shouldReceive('expiresAfter')->once()->with(null);
-        $this->itemMock->shouldReceive('expiresAfter')->once()->with(2);
-        $this->mock->shouldReceive('getItem')->twice()->with('some_item')->andReturn($this->itemMock);
-        $this->mock->shouldReceive('save')->twice()->with($this->itemMock)->andReturn(true);
+        $this
+            ->itemMock
+            ->method('set')
+            ->with('dummy_data')
+            ->willReturnSelf();
+        $this
+            ->itemMock
+            ->method('expiresAfter')
+            ->willReturnMap([
+                [null, $this->itemMock],
+                [2, $this->itemMock],
+            ]);
 
-        $this->assertTrue($this->bridge->set('some_item', 'dummy_data'));
-        $this->assertTrue($this->bridge->set('some_item', 'dummy_data', 2));
+        $this
+            ->poolMock
+            ->method('getItem')
+            ->with('some_item')
+            ->willReturn($this->itemMock);
+        $this
+            ->poolMock
+            ->method('save')
+            ->with($this->itemMock)
+            ->willReturn(true);
+
+        static::assertTrue($this->bridge->set('some_item', 'dummy_data'));
+        static::assertTrue($this->bridge->set('some_item', 'dummy_data', 2));
     }
 
-    public function testDelete()
+    public function testDelete(): void
     {
-        $this->mock->shouldReceive('deleteItem')->once()->with('some_item')->andReturn(true);
+        $this
+            ->poolMock
+            ->expects(static::once())
+            ->method('deleteItem')
+            ->with('some_item')
+            ->willReturn(true);
 
-        $this->assertTrue($this->bridge->delete('some_item'));
+        static::assertTrue($this->bridge->delete('some_item'));
     }
 }
